Clean test base
