+++ systemtests/src/main/java/io/enmasse/systemtest/listener/JunitCallbackListener.java
@@ -231,9 +231,9 @@ public class JunitCallbackListener implements TestExecutionExceptionHandler, Lif
             Files.writeString(path.resolve("describe_nodes.txt"), KubeCMDClient.describeNodes().getStdOut());
             Files.writeString(path.resolve("events.txt"), KubeCMDClient.getEvents(kube.getInfraNamespace()).getStdOut());
             Files.writeString(path.resolve("configmaps.yaml"), KubeCMDClient.getConfigmaps(kube.getInfraNamespace()).getStdOut());
-            Files.writeString(path.resolve("pvs.txt"), KubeCMDClient.runOnCluster("describe", "pv").getStdOut());
-            Files.writeString(path.resolve("pvcs.txt"), KubeCMDClient.runOnCluster("describe", "pvc", "-n", Kubernetes.getInstance().getInfraNamespace()).getStdOut());
-            Files.writeString(path.resolve("storageclass.yml"), KubeCMDClient.runOnCluster("get", "storageclass", "-o", "yaml").getStdOut());
+            Files.writeString(path.resolve("pvs.txt"), KubeCMDClient.runOnClusterWithoutLogger("describe", "pv").getStdOut());
+            Files.writeString(path.resolve("pvcs.txt"), KubeCMDClient.runOnClusterWithoutLogger("describe", "pvc", "-n", Kubernetes.getInstance().getInfraNamespace()).getStdOut());
+            Files.writeString(path.resolve("storageclass.yml"), KubeCMDClient.runOnClusterWithoutLogger("get", "storageclass", "-o", "yaml").getStdOut());
             if (testInfo.isClassIoT()) {
                 Files.writeString(path.resolve("iotconfig.yaml"), KubeCMDClient.getIoTConfig(kube.getInfraNamespace()).getStdOut());
                 GlobalLogCollector collectors = new GlobalLogCollector(kube, path, kube.getInfraNamespace());
