+++ systemtests/src/test/java/io/enmasse/systemtest/bases/TestBase.java
@@ -9,26 +9,16 @@ import com.google.common.collect.Ordering;
 import io.enmasse.address.model.Address;
 import io.enmasse.address.model.AddressBuilder;
 import io.enmasse.address.model.AddressSpace;
-import io.enmasse.address.model.AddressSpaceSchemaList;
-import io.enmasse.systemtest.Endpoint;
 import io.enmasse.systemtest.Environment;
 import io.enmasse.systemtest.IndicativeSentences;
 import io.enmasse.systemtest.UserCredentials;
 import io.enmasse.systemtest.amqp.AmqpClient;
 import io.enmasse.systemtest.broker.BrokerManagement;
 import io.enmasse.systemtest.info.TestInfo;
-import io.enmasse.systemtest.iot.HttpAdapterClient;
-import io.enmasse.systemtest.iot.MessageType;
 import io.enmasse.systemtest.listener.JunitCallbackListener;
 import io.enmasse.systemtest.logs.GlobalLogCollector;
 import io.enmasse.systemtest.manager.ResourceManager;
-import io.enmasse.systemtest.messagingclients.ClientArgument;
-import io.enmasse.systemtest.messagingclients.ExternalMessagingClient;
-import io.enmasse.systemtest.messagingclients.rhea.RheaClientConnector;
-import io.enmasse.systemtest.messagingclients.rhea.RheaClientReceiver;
-import io.enmasse.systemtest.messagingclients.rhea.RheaClientSender;
 import io.enmasse.systemtest.model.address.AddressType;
-import io.enmasse.systemtest.model.addressplan.DestinationPlan;
 import io.enmasse.systemtest.mqtt.MqttUtils;
 import io.enmasse.systemtest.platform.KubeCMDClient;
 import io.enmasse.systemtest.selenium.SeleniumManagement;
@@ -37,12 +27,10 @@ import io.enmasse.systemtest.selenium.page.ConsoleWebPage;
 import io.enmasse.systemtest.time.TimeoutBudget;
 import io.enmasse.systemtest.utils.AddressSpaceUtils;
 import io.enmasse.systemtest.utils.AddressUtils;
-import io.enmasse.systemtest.utils.IoTUtils;
 import io.enmasse.systemtest.utils.JmsProvider;
 import io.enmasse.systemtest.utils.TestUtils;
 import io.enmasse.systemtest.utils.UserUtils;
 import io.enmasse.user.model.v1.Operation;
-import io.enmasse.user.model.v1.User;
 import io.enmasse.user.model.v1.UserAuthorizationBuilder;
 import org.apache.qpid.proton.amqp.messaging.AmqpValue;
 import org.apache.qpid.proton.message.Message;
@@ -120,82 +108,14 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
     //======================================= Help methods ===========================================
     //================================================================================================
 
-    protected boolean userExist(AddressSpace addressSpace, String username) {
-        return resourcesManager.getUser(addressSpace, username) != null;
-    }
-
-    /**
-     * give you a schema object
-     *
-     * @return schema object
-     * @throws Exception
-     */
-    protected AddressSpaceSchemaList getSchema() throws Exception {
-        return kubernetes.getSchemaClient().list();
-    }
-
-    protected Endpoint getMessagingRoute(AddressSpace addressSpace) throws Exception {
-        Endpoint messagingEndpoint = AddressSpaceUtils.getEndpointByServiceName(addressSpace, "messaging");
-        if (messagingEndpoint == null) {
-            String externalEndpointName = AddressSpaceUtils.getExternalEndpointName(addressSpace, "messaging-" + AddressSpaceUtils.getAddressSpaceInfraUuid(addressSpace));
-            messagingEndpoint = kubernetes.getExternalEndpoint(externalEndpointName);
-        }
-        if (TestUtils.resolvable(messagingEndpoint)) {
-            return messagingEndpoint;
-        } else {
-            return kubernetes.getEndpoint("messaging-" + AddressSpaceUtils.getAddressSpaceInfraUuid(addressSpace), addressSpace.getMetadata().getNamespace(), "amqps");
-        }
-    }
-
-    protected String getOCConsoleRoute() {
-        if (kubernetes.getOcpVersion() == 4) {
-            return String.format("https://console-openshift-console.%s", environment.kubernetesDomain()).replaceAll("(?<!(http:|https:))[//]+", "/");
-        } else {
-            return String.format("%s/console", environment.getApiUrl()).replaceAll("(?<!(http:|https:))[//]+", "/");
-        }
-    }
-
-    protected String getConsoleRoute(AddressSpace addressSpace) {
-        Endpoint consoleEndpoint = getConsoleEndpoint(addressSpace);
-        String consoleRoute = String.format("https://%s", consoleEndpoint.toString());
-        LOGGER.info(consoleRoute);
-        return consoleRoute;
-    }
-
-    protected Endpoint getConsoleEndpoint(AddressSpace addressSpace) {
-        Endpoint consoleEndpoint = AddressSpaceUtils.getEndpointByServiceName(addressSpace, "console");
-        if (consoleEndpoint == null) {
-            String externalEndpointName = AddressSpaceUtils.getExternalEndpointName(addressSpace, "console");
-            consoleEndpoint = kubernetes.getExternalEndpoint(externalEndpointName);
-        }
-        return consoleEndpoint;
-    }
-
-    /**
-     * selenium provider with Firefox webdriver
-     */
-    protected SeleniumProvider getFirefoxSeleniumProvider() throws Exception {
-        SeleniumProvider seleniumProvider = SeleniumProvider.getInstance();
-        seleniumProvider.setupDriver(TestUtils.getFirefoxDriver());
-        return seleniumProvider;
-    }
-
-    protected void waitForSubscribersConsole(AddressSpace addressSpace, Address destination, int expectedCount) throws Exception {
-        int budget = 60; //seconds
-        waitForSubscribersConsole(addressSpace, destination, expectedCount, budget);
-    }
-
-    /**
-     * wait for expected count of subscribers on topic (check via console)
-     *
-     * @param budget timeout budget in seconds
-     */
-    private void waitForSubscribersConsole(AddressSpace addressSpace, Address destination, int expectedCount, int budget) throws Exception {
+    public void waitForSubscribersConsole(AddressSpace addressSpace, Address destination, int expectedCount) throws Exception {
+        int budget = 60;
         SeleniumProvider selenium = null;
         try {
             SeleniumManagement.deployFirefoxApp();
-            selenium = getFirefoxSeleniumProvider();
-            ConsoleWebPage console = new ConsoleWebPage(selenium, getConsoleRoute(addressSpace), addressSpace, clusterUser);
+            selenium = SeleniumProvider.getInstance();
+            selenium.setupDriver(TestUtils.getFirefoxDriver());
+            ConsoleWebPage console = new ConsoleWebPage(selenium, AddressSpaceUtils.getConsoleRoute(addressSpace), addressSpace, clusterUser);
             console.openWebConsolePage();
             console.openAddressesPageWebConsole();
 
@@ -210,20 +130,8 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
         }
     }
 
-    /**
-     * wait for expected count of subscribers on topic
-     *
-     * @param addressSpace
-     * @param topic         name of topic
-     * @param expectedCount count of expected subscribers
-     * @throws Exception
-     */
-    protected void waitForSubscribers(BrokerManagement brokerManagement, AddressSpace addressSpace, String topic, int expectedCount) throws Exception {
+    public void waitForSubscribers(BrokerManagement brokerManagement, AddressSpace addressSpace, String topic, int expectedCount) throws Exception {
         TimeoutBudget budget = new TimeoutBudget(1, TimeUnit.MINUTES);
-        waitForSubscribers(brokerManagement, addressSpace, topic, expectedCount, budget);
-    }
-
-    private void waitForSubscribers(BrokerManagement brokerManagement, AddressSpace addressSpace, String topic, int expectedCount, TimeoutBudget budget) throws Exception {
         AmqpClient queueClient = null;
         try {
             queueClient = resourcesManager.getAmqpClientFactory().createQueueClient(addressSpace);
@@ -243,7 +151,7 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
             resourcesManager.appendAddresses(replyQueue);
 
             boolean done = false;
-            int actualSubscribers = 0;
+            int actualSubscribers;
             do {
                 actualSubscribers = getSubscriberCount(brokerManagement, queueClient, replyQueue, topic);
                 LOGGER.info("Have " + actualSubscribers + " subscribers. Expecting " + expectedCount);
@@ -261,19 +169,25 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
         }
     }
 
-    private void waitForBrokerReplicas(AddressSpace addressSpace, Address destination, int expectedReplicas, boolean readyRequired, TimeoutBudget budget, long checkInterval) throws Exception {
-        TestUtils.waitForNBrokerReplicas(addressSpace, expectedReplicas, readyRequired, destination, budget, checkInterval);
-    }
+    private int getSubscriberCount(BrokerManagement brokerManagement, AmqpClient queueClient, Address
+            replyQueue, String topic) throws Exception {
+        List<String> queueNames = brokerManagement.getQueueNames(queueClient, replyQueue, topic);
 
-    private void waitForBrokerReplicas(AddressSpace addressSpace, Address destination,
-                                       int expectedReplicas, TimeoutBudget budget) throws Exception {
-        waitForBrokerReplicas(addressSpace, destination, expectedReplicas, true, budget, 5000);
+        AtomicInteger subscriberCount = new AtomicInteger(0);
+        queueNames.forEach((String queue) -> {
+            try {
+                subscriberCount.addAndGet(brokerManagement.getSubscriberCount(queueClient, replyQueue, queue));
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        });
+        return subscriberCount.get();
     }
 
     protected void waitForBrokerReplicas(AddressSpace addressSpace, Address destination, int expectedReplicas) throws
             Exception {
         TimeoutBudget budget = new TimeoutBudget(10, TimeUnit.MINUTES);
-        waitForBrokerReplicas(addressSpace, destination, expectedReplicas, budget);
+        TestUtils.waitForNBrokerReplicas(addressSpace, expectedReplicas, true, destination, budget, 5000);
     }
 
     protected void waitForRouterReplicas(AddressSpace addressSpace, int expectedReplicas) throws
@@ -291,301 +205,6 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
                 .noneMatch(pod -> uids.contains(pod.getMetadata().getUid()))), new TimeoutBudget(2, TimeUnit.MINUTES));
     }
 
-    /**
-     * Wait for destinations are in isReady=true state within default timeout (10 MINUTE)
-     */
-    protected void waitForDestinationsReady(Address... destinations) throws Exception {
-        TimeoutBudget budget = new TimeoutBudget(10, TimeUnit.MINUTES);
-        AddressUtils.waitForDestinationsReady(budget, destinations);
-    }
-
-    /**
-     * return list of queue names created for subscribers
-     *
-     * @param queueClient
-     * @param replyQueue  queue for answer is required
-     * @param topic       topic name
-     * @return
-     * @throws Exception
-     */
-    private List<String> getBrokerQueueNames(BrokerManagement brokerManagement, AmqpClient
-            queueClient, Address replyQueue, String topic) throws Exception {
-        return brokerManagement.getQueueNames(queueClient, replyQueue, topic);
-    }
-
-    /**
-     * get count of subscribers subscribed to 'topic'
-     *
-     * @param queueClient queue client with admin permissions
-     * @param replyQueue  queue for answer is required
-     * @param topic       topic name
-     * @return
-     * @throws Exception
-     */
-    private int getSubscriberCount(BrokerManagement brokerManagement, AmqpClient queueClient, Address
-            replyQueue, String topic) throws Exception {
-        List<String> queueNames = getBrokerQueueNames(brokerManagement, queueClient, replyQueue, topic);
-
-        AtomicInteger subscriberCount = new AtomicInteger(0);
-        queueNames.forEach((String queue) -> {
-            try {
-                subscriberCount.addAndGet(brokerManagement.getSubscriberCount(queueClient, replyQueue, queue));
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        });
-        return subscriberCount.get();
-    }
-
-    protected ArrayList<Address> generateQueueTopicList(AddressSpace addressspace, String infix, IntStream range) {
-        ArrayList<Address> addresses = new ArrayList<>();
-        range.forEach(i -> {
-            if (i % 2 == 0) {
-                addresses.add(new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, String.format("topic-%s-%d", infix, i)))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("topic")
-                        .withAddress(String.format("topic-%s-%d", infix, i))
-                        .withPlan(getDefaultPlan(AddressType.TOPIC))
-                        .endSpec()
-                        .build());
-            } else {
-                addresses.add(new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, String.format("queue-%s-%d", infix, i)))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("queue")
-                        .withAddress(String.format("queue-%s-%d", infix, i))
-                        .withPlan(getDefaultPlan(AddressType.QUEUE))
-                        .endSpec()
-                        .build());
-            }
-        });
-        return addresses;
-    }
-
-    /**
-     * attach N receivers into one address with own username/password
-     */
-    public List<ExternalMessagingClient> attachReceivers(AddressSpace addressSpace, Address destination,
-                                                int receiverCount, int timeout, UserCredentials userCredentials) throws Exception {
-
-        List<ExternalMessagingClient> receivers = new ArrayList<>();
-        for (int i = 0; i < receiverCount; i++) {
-            ExternalMessagingClient receiverClient = new ExternalMessagingClient()
-                    .withClientEngine(new RheaClientReceiver())
-                    .withMessagingRoute(getMessagingRoute(addressSpace))
-                    .withAddress(destination)
-                    .withCredentials(userCredentials)
-                    .withTimeout(timeout)
-                    .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property1~50")
-                    .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property2~testValue");
-            receiverClient.runAsync(false);
-            receivers.add(receiverClient);
-        }
-
-        Thread.sleep(15000); //wait for attached
-        return receivers;
-    }
-
-    /**
-     * attach senders to destinations (for N-th destination is attached N+1 senders)
-     */
-    public List<ExternalMessagingClient> attachSenders(AddressSpace addressSpace, List<Address> destinations, int timeout, UserCredentials userCredentials) throws Exception {
-        List<ExternalMessagingClient> senders = new ArrayList<>();
-
-        for (int i = 0; i < destinations.size(); i++) {
-            for (int j = 0; j < i + 1; j++) {
-                ExternalMessagingClient senderClient = new ExternalMessagingClient()
-                        .withClientEngine(new RheaClientSender())
-                        .withMessagingRoute(getMessagingRoute(addressSpace))
-                        .withAddress(destinations.get(i))
-                        .withCredentials(userCredentials)
-                        .withMessageBody("msg no.%d")
-                        .withTimeout(timeout)
-                        .withCount(30)
-                        .withAdditionalArgument(ClientArgument.DURATION, 30)
-                        .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property1~50")
-                        .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property2~testValue");
-                senderClient.runAsync(false);
-                senders.add(senderClient);
-            }
-        }
-
-        return senders;
-    }
-
-    /**
-     * attach receivers to destinations (for N-th destination is attached N+1 senders)
-     */
-    public List<ExternalMessagingClient> attachReceivers(AddressSpace addressSpace, List<Address> destinations, int timeout, UserCredentials userCredentials) throws Exception {
-        List<ExternalMessagingClient> receivers = new ArrayList<>();
-
-        for (int i = 0; i < destinations.size(); i++) {
-            for (int j = 0; j < i + 1; j++) {
-                ExternalMessagingClient receiverClient = new ExternalMessagingClient()
-                        .withClientEngine(new RheaClientReceiver())
-                        .withMessagingRoute(getMessagingRoute(addressSpace))
-                        .withAddress(destinations.get(i))
-                        .withCredentials(userCredentials)
-                        .withTimeout(timeout)
-                        .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property1~50")
-                        .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property2~testValue");
-                receiverClient.runAsync(false);
-                receivers.add(receiverClient);
-            }
-        }
-
-        return receivers;
-    }
-
-    /**
-     * create M connections with N receivers and K senders
-     */
-    protected ExternalMessagingClient attachConnector(AddressSpace addressSpace, Address destination,
-                                             int connectionCount,
-                                             int senderCount, int receiverCount, UserCredentials credentials, int timeout) throws Exception {
-
-        ExternalMessagingClient connectorClient = new ExternalMessagingClient()
-                .withClientEngine(new RheaClientConnector())
-                .withMessagingRoute(getMessagingRoute(addressSpace))
-                .withAddress(destination)
-                .withCredentials(credentials)
-                .withCount(connectionCount)
-                .withTimeout(timeout)
-                .withAdditionalArgument(ClientArgument.SENDER_COUNT, Integer.toString(senderCount))
-                .withAdditionalArgument(ClientArgument.RECEIVER_COUNT, Integer.toString(receiverCount))
-                .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property1~50")
-                .withAdditionalArgument(ClientArgument.CONN_PROPERTY, "connection_property2~testValue");
-        connectorClient.runAsync(false);
-
-        return connectorClient;
-    }
-
-    /**
-     * stop all clients from list of Abstract clients
-     */
-    protected void stopClients(List<ExternalMessagingClient> clients) {
-        if (clients != null) {
-            LOGGER.info("Stopping clients...");
-            clients.forEach(ExternalMessagingClient::stop);
-        }
-    }
-
-    /**
-     * create users and groups for wildcard authz tests
-     */
-    protected List<User> createUsersWildcard(AddressSpace addressSpace, Operation operation) throws
-            Exception {
-        List<User> users = new ArrayList<>();
-        users.add(UserUtils.createUserResource(new UserCredentials("user1", "password"))
-                .editSpec()
-                .withAuthorization(Collections.singletonList(new UserAuthorizationBuilder()
-                        .withAddresses("*")
-                        .withOperations(operation)
-                        .build()))
-                .endSpec()
-                .done());
-
-        users.add(UserUtils.createUserResource(new UserCredentials("user2", "password"))
-                .editSpec()
-                .withAuthorization(Collections.singletonList(new UserAuthorizationBuilder()
-                        .withAddresses("queue/*")
-                        .withOperations(operation)
-                        .build()))
-                .endSpec()
-                .done());
-
-        users.add(UserUtils.createUserResource(new UserCredentials("user3", "password"))
-                .editSpec()
-                .withAuthorization(Collections.singletonList(new UserAuthorizationBuilder()
-                        .withAddresses("topic/*")
-                        .withOperations(operation)
-                        .build()))
-                .endSpec()
-                .done());
-
-        users.add(UserUtils.createUserResource(new UserCredentials("user4", "password"))
-                .editSpec()
-                .withAuthorization(Collections.singletonList(new UserAuthorizationBuilder()
-                        .withAddresses("queueA*")
-                        .withOperations(operation)
-                        .build()))
-                .endSpec()
-                .done());
-
-        users.add(UserUtils.createUserResource(new UserCredentials("user5", "password"))
-                .editSpec()
-                .withAuthorization(Collections.singletonList(new UserAuthorizationBuilder()
-                        .withAddresses("topicA*")
-                        .withOperations(operation)
-                        .build()))
-                .endSpec()
-                .done());
-
-        for (User user : users) {
-            resourcesManager.createOrUpdateUser(addressSpace, user);
-        }
-        return users;
-    }
-
-    protected List<Address> getAddressesWildcard(AddressSpace addressspace) {
-        Address queue = new AddressBuilder()
-                .withNewMetadata()
-                .withNamespace(addressspace.getMetadata().getNamespace())
-                .withName(AddressUtils.generateAddressMetadataName(addressspace, "queue/1234"))
-                .endMetadata()
-                .withNewSpec()
-                .withType("queue")
-                .withAddress("queue/1234")
-                .withPlan(getDefaultPlan(AddressType.QUEUE))
-                .endSpec()
-                .build();
-
-        Address queue2 = new AddressBuilder()
-                .withNewMetadata()
-                .withNamespace(addressspace.getMetadata().getNamespace())
-                .withName(AddressUtils.generateAddressMetadataName(addressspace, "queue/ABCD"))
-                .endMetadata()
-                .withNewSpec()
-                .withType("queue")
-                .withAddress("queue/ABCD")
-                .withPlan(getDefaultPlan(AddressType.QUEUE))
-                .endSpec()
-                .build();
-
-        Address topic = new AddressBuilder()
-                .withNewMetadata()
-                .withNamespace(addressspace.getMetadata().getNamespace())
-                .withName(AddressUtils.generateAddressMetadataName(addressspace, "topic/2345"))
-                .endMetadata()
-                .withNewSpec()
-                .withType("topic")
-                .withAddress("topic/2345")
-                .withPlan(getDefaultPlan(AddressType.TOPIC))
-                .endSpec()
-                .build();
-
-        Address topic2 = new AddressBuilder()
-                .withNewMetadata()
-                .withNamespace(addressspace.getMetadata().getNamespace())
-                .withName(AddressUtils.generateAddressMetadataName(addressspace, "topic/ABCD"))
-                .endMetadata()
-                .withNewSpec()
-                .withType("topic")
-                .withAddress("topic/ABCD")
-                .withPlan(getDefaultPlan(AddressType.TOPIC))
-                .endSpec()
-                .build();
-
-        return Arrays.asList(queue, queue2, topic, topic2);
-    }
-
     protected void logWithSeparator(Logger logger, String... messages) {
         logger.info("--------------------------------------------------------------------------------");
         for (String message : messages) {
@@ -593,201 +212,15 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
         }
     }
 
-    /**
-     * body for rest api tests
-     */
-    protected void runRestApiTest(AddressSpace addressSpace, Address d1, Address d2) throws Exception {
-        List<String> destinationsNames = Arrays.asList(d1.getSpec().getAddress(), d2.getSpec().getAddress());
-        resourcesManager.setAddresses(d1);
-        resourcesManager.appendAddresses(d2);
-
-        //d1, d2
-        List<String> response = AddressUtils.getAddresses(addressSpace).stream().map(address -> address.getSpec().getAddress()).collect(Collectors.toList());
-        assertThat("Rest api does not return all addresses", response, is(destinationsNames));
-        LOGGER.info("addresses {} successfully created", Arrays.toString(destinationsNames.toArray()));
-
-        //get specific address d2
-        Address res = kubernetes.getAddressClient(addressSpace.getMetadata().getNamespace()).withName(d2.getMetadata().getName()).get();
-        assertThat("Rest api does not return specific address", res.getSpec().getAddress(), is(d2.getSpec().getAddress()));
-
-        resourcesManager.deleteAddresses(d1);
-
-        //d2
-        response = AddressUtils.getAddresses(addressSpace).stream().map(address -> address.getSpec().getAddress()).collect(Collectors.toList());
-        assertThat("Rest api does not return right addresses", response, is(destinationsNames.subList(1, 2)));
-        LOGGER.info("address {} successfully deleted", d1.getSpec().getAddress());
-
-        resourcesManager.deleteAddresses(d2);
-
-        //empty
-        List<Address> listRes = AddressUtils.getAddresses(addressSpace);
-        assertThat("Rest api returns addresses", listRes, is(Collections.emptyList()));
-        LOGGER.info("addresses {} successfully deleted", d2.getSpec().getAddress());
-
-        resourcesManager.setAddresses(d1, d2);
-        resourcesManager.deleteAddresses(d1, d2);
-
-        listRes = AddressUtils.getAddresses(addressSpace);
-        assertThat("Rest api returns addresses", listRes, is(Collections.emptyList()));
-        LOGGER.info("addresses {} successfully deleted", Arrays.toString(destinationsNames.toArray()));
-    }
-
-    protected void sendReceiveLargeMessageQueue(JmsProvider jmsProvider, double sizeInMB, Address dest, int count) throws Exception {
-        sendReceiveLargeMessageQueue(jmsProvider, sizeInMB, dest, count, DeliveryMode.NON_PERSISTENT);
-    }
-
-    protected void sendReceiveLargeMessageQueue(JmsProvider jmsProvider, double sizeInMB, Address dest, int count, int mode) throws Exception {
-        int size = (int) (sizeInMB * 1024 * 1024);
-
-        Session session = jmsProvider.getConnection().createSession(false, Session.AUTO_ACKNOWLEDGE);
-        javax.jms.Queue testQueue = (javax.jms.Queue) jmsProvider.getDestination(dest.getSpec().getAddress());
-        List<javax.jms.Message> messages = jmsProvider.generateMessages(session, count, size);
-
-        MessageProducer sender = session.createProducer(testQueue);
-        MessageConsumer receiver = session.createConsumer(testQueue);
-
-        sendReceiveLargeMessage(jmsProvider, sender, receiver, sizeInMB, mode, count, messages);
-
-    }
-
-    protected void sendReceiveLargeMessageTopic(JmsProvider jmsProvider, double sizeInMB, Address dest, int count) throws Exception {
-        sendReceiveLargeMessageTopic(jmsProvider, sizeInMB, dest, count, DeliveryMode.NON_PERSISTENT);
-    }
-
-    protected void sendReceiveLargeMessageTopic(JmsProvider jmsProvider, double sizeInMB, Address dest, int count, int mode) throws Exception {
-        int size = (int) (sizeInMB * 1024 * 1024);
-
-        Session session = jmsProvider.getConnection().createSession(false, Session.AUTO_ACKNOWLEDGE);
-        javax.jms.Topic testTopic = (javax.jms.Topic) jmsProvider.getDestination(dest.getSpec().getAddress());
-        List<javax.jms.Message> messages = jmsProvider.generateMessages(session, count, size);
-
-        MessageProducer sender = session.createProducer(testTopic);
-        MessageConsumer receiver = session.createConsumer(testTopic);
-
-        sendReceiveLargeMessage(jmsProvider, sender, receiver, sizeInMB, mode, count, messages);
-        session.close();
-        sender.close();
-        receiver.close();
-    }
-
-    private void sendReceiveLargeMessage(JmsProvider jmsProvider, MessageProducer sender, MessageConsumer receiver, double sizeInMB, int mode, int count, List<javax.jms.Message> messages) {
-        List<javax.jms.Message> recvd;
-
-        jmsProvider.sendMessages(sender, messages, mode, javax.jms.Message.DEFAULT_PRIORITY, javax.jms.Message.DEFAULT_TIME_TO_LIVE);
-        LOGGER.info("{}MB {} message sent", sizeInMB, mode == DeliveryMode.PERSISTENT ? "durable" : "non-durable");
-
-        recvd = jmsProvider.receiveMessages(receiver, count, 2000);
-        assertThat("Wrong count of received messages", recvd.size(), Matchers.is(count));
-        LOGGER.info("{}MB {} message received", sizeInMB, mode == DeliveryMode.PERSISTENT ? "durable" : "non-durable");
-    }
-
-    protected List<Address> getAllStandardAddresses(AddressSpace addressspace) {
-        return Arrays.asList(
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-queue"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("queue")
-                        .withAddress("test-queue")
-                        .withPlan(DestinationPlan.STANDARD_SMALL_QUEUE)
-                        .endSpec()
-                        .build(),
-
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-topic"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("topic")
-                        .withAddress("test-topic")
-                        .withPlan(DestinationPlan.STANDARD_SMALL_TOPIC)
-                        .endSpec()
-                        .build(),
-
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-queue-sharded"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("queue")
-                        .withAddress("test-queue-sharded")
-                        .withPlan(DestinationPlan.STANDARD_LARGE_QUEUE)
-                        .endSpec()
-                        .build(),
-
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-topic-sharded"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("topic")
-                        .withAddress("test-topic-sharded")
-                        .withPlan(DestinationPlan.STANDARD_LARGE_TOPIC)
-                        .endSpec()
-                        .build(),
-
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-anycast"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("anycast")
-                        .withAddress("test-anycast")
-                        .withPlan(DestinationPlan.STANDARD_SMALL_ANYCAST)
-                        .endSpec()
-                        .build(),
-
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-multicast"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("multicast")
-                        .withAddress("test-multicast")
-                        .withPlan(DestinationPlan.STANDARD_SMALL_MULTICAST)
-                        .endSpec()
-                        .build());
-    }
-
-    protected List<Address> getAllBrokeredAddresses(AddressSpace addressspace) {
-        return Arrays.asList(
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-queue"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("queue")
-                        .withAddress("test-queue")
-                        .withPlan(DestinationPlan.BROKERED_QUEUE)
-                        .endSpec()
-                        .build(),
-
-                new AddressBuilder()
-                        .withNewMetadata()
-                        .withNamespace(addressspace.getMetadata().getNamespace())
-                        .withName(AddressUtils.generateAddressMetadataName(addressspace, "test-topic"))
-                        .endMetadata()
-                        .withNewSpec()
-                        .withType("topic")
-                        .withAddress("test-topic")
-                        .withPlan(DestinationPlan.BROKERED_TOPIC)
-                        .endSpec()
-                        .build());
-    }
-
     protected List<String> extractBodyAsString(Future<List<Message>> msgs) throws Exception {
         return msgs.get(1, TimeUnit.MINUTES).stream().map(m -> (String) ((AmqpValue) m.getBody()).getValue()).collect(Collectors.toList());
     }
 
-    protected static void simpleMQTTSendReceive(Address dest, IMqttClient client, int msgCount) throws Exception {
+    //================================================================================================
+    //==================================== Asserts methods ===========================================
+    //================================================================================================
+
+    protected static void assertSimpleMQTTSendReceive(Address dest, IMqttClient client, int msgCount) throws Exception {
         List<MqttMessage> messages = IntStream.range(0, msgCount).boxed().map(i -> {
             MqttMessage m = new MqttMessage();
             m.setPayload(String.format("mqtt-simple-send-receive-%s", i).getBytes(StandardCharsets.UTF_8));
@@ -807,9 +240,6 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
                 receivedCount, is(messages.size()));
     }
 
-    //================================================================================================
-    //==================================== Asserts methods ===========================================
-    //================================================================================================
     protected <T extends Comparable<T>> void assertSorted(String message, Iterable<T> list) throws Exception {
         assertSorted(message, list, false);
     }
@@ -849,26 +279,13 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
         fail(String.format("Incorrect result value! expected: '%s', got: '%s'", expected, Objects.requireNonNull(got)));
     }
 
-    protected <T> void assertWaitForValue(T expected, Callable<T> fn) throws Exception {
-        assertWaitForValue(expected, fn, new TimeoutBudget(10, TimeUnit.SECONDS));
+    protected static void assertDefaultEnabled(final Boolean enabled) {
+        if (enabled != null && !Boolean.TRUE.equals(enabled)) {
+            fail("Default value must be 'null' or 'true'");
+        }
     }
 
-    //================================================================================================
-    //====================================== Shared test help methods =======================================
-    //================================================================================================
-
-    /**
-     * Create users within groups (according to destNamePrefix and customerIndex), wait until destinations are ready to use
-     * and start sending and receiving messages
-     *
-     * @param dest           list of all available destinations (destinations are not in ready state presumably)
-     * @param users          list of users dedicated for sending messages into destinations above
-     * @param destNamePrefix prefix of destinations name (due to authorization)
-     * @param customerIndex  also important due to authorization (only users under this customer can send messages into dest)
-     * @param messageCount   count of messages that will be send into destinations
-     * @throws Exception
-     */
-    protected void doMessaging(List<Address> dest, List<UserCredentials> users, String destNamePrefix, int customerIndex, int messageCount) throws Exception {
+    protected void assertConcurentMessaging(List<Address> dest, List<UserCredentials> users, String destNamePrefix, int customerIndex, int messageCount) throws Exception {
         ArrayList<AmqpClient> clients = new ArrayList<>(users.size());
         String sufix = AddressSpaceUtils.isBrokered(resourcesManager.getSharedAddressSpace()) ? "#" : "*";
         users.forEach((user) -> {
@@ -890,7 +307,7 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
             }
         });
 
-        waitForDestinationsReady(dest.toArray(new Address[0]));
+        AddressUtils.waitForDestinationsReady(dest.toArray(new Address[0]));
         //start sending messages
         int everyN = 3;
         for (AmqpClient client : clients) {
@@ -918,24 +335,88 @@ public abstract class TestBase implements ITestBase, ITestSeparator {
         }
     }
 
-    //================================================================================================
-    //====================================== IoT test methods  =======================================
-    //================================================================================================
-
-    protected void waitForFirstSuccessOnTelemetry(HttpAdapterClient adapterClient) throws Exception {
-        IoTUtils.waitForFirstSuccess(adapterClient, MessageType.TELEMETRY);
+    protected void assertSendReceiveLargeMessageQueue(JmsProvider jmsProvider, double sizeInMB, Address dest, int count) throws Exception {
+        assertSendReceiveLargeMessageQueue(jmsProvider, sizeInMB, dest, count, DeliveryMode.NON_PERSISTENT);
     }
 
-    /**
-     * Test if the enabled flag is set to "enabled".
-     * <br>
-     * The flag is considered "enabled", in case the value is "true" or missing.
-     *
-     * @param enabled The object to test.
-     */
-    protected static void assertDefaultEnabled(final Boolean enabled) {
-        if (enabled != null && !Boolean.TRUE.equals(enabled)) {
-            fail("Default value must be 'null' or 'true'");
-        }
+    protected void assertSendReceiveLargeMessageQueue(JmsProvider jmsProvider, double sizeInMB, Address dest, int count, int mode) throws Exception {
+        int size = (int) (sizeInMB * 1024 * 1024);
+
+        Session session = jmsProvider.getConnection().createSession(false, Session.AUTO_ACKNOWLEDGE);
+        javax.jms.Queue testQueue = (javax.jms.Queue) jmsProvider.getDestination(dest.getSpec().getAddress());
+        List<javax.jms.Message> messages = jmsProvider.generateMessages(session, count, size);
+
+        MessageProducer sender = session.createProducer(testQueue);
+        MessageConsumer receiver = session.createConsumer(testQueue);
+
+        assertSendReceiveLargeMessage(jmsProvider, sender, receiver, sizeInMB, mode, count, messages);
+
+    }
+
+    protected void assertSendReceiveLargeMessageTopic(JmsProvider jmsProvider, double sizeInMB, Address dest, int count) throws Exception {
+        assertSendReceiveLargeMessageTopic(jmsProvider, sizeInMB, dest, count, DeliveryMode.NON_PERSISTENT);
+    }
+
+    protected void assertSendReceiveLargeMessageTopic(JmsProvider jmsProvider, double sizeInMB, Address dest, int count, int mode) throws Exception {
+        int size = (int) (sizeInMB * 1024 * 1024);
+
+        Session session = jmsProvider.getConnection().createSession(false, Session.AUTO_ACKNOWLEDGE);
+        javax.jms.Topic testTopic = (javax.jms.Topic) jmsProvider.getDestination(dest.getSpec().getAddress());
+        List<javax.jms.Message> messages = jmsProvider.generateMessages(session, count, size);
+
+        MessageProducer sender = session.createProducer(testTopic);
+        MessageConsumer receiver = session.createConsumer(testTopic);
+
+        assertSendReceiveLargeMessage(jmsProvider, sender, receiver, sizeInMB, mode, count, messages);
+        session.close();
+        sender.close();
+        receiver.close();
+    }
+
+    private void assertSendReceiveLargeMessage(JmsProvider jmsProvider, MessageProducer sender, MessageConsumer receiver, double sizeInMB, int mode, int count, List<javax.jms.Message> messages) {
+        List<javax.jms.Message> recvd;
+
+        jmsProvider.sendMessages(sender, messages, mode, javax.jms.Message.DEFAULT_PRIORITY, javax.jms.Message.DEFAULT_TIME_TO_LIVE);
+        LOGGER.info("{}MB {} message sent", sizeInMB, mode == DeliveryMode.PERSISTENT ? "durable" : "non-durable");
+
+        recvd = jmsProvider.receiveMessages(receiver, count, 2000);
+        assertThat("Wrong count of received messages", recvd.size(), Matchers.is(count));
+        LOGGER.info("{}MB {} message received", sizeInMB, mode == DeliveryMode.PERSISTENT ? "durable" : "non-durable");
+    }
+
+    protected void assertAddressApi(AddressSpace addressSpace, Address d1, Address d2) throws Exception {
+        List<String> destinationsNames = Arrays.asList(d1.getSpec().getAddress(), d2.getSpec().getAddress());
+        resourcesManager.setAddresses(d1);
+        resourcesManager.appendAddresses(d2);
+
+        //d1, d2
+        List<String> response = AddressUtils.getAddresses(addressSpace).stream().map(address -> address.getSpec().getAddress()).collect(Collectors.toList());
+        assertThat("Rest api does not return all addresses", response, is(destinationsNames));
+        LOGGER.info("addresses {} successfully created", Arrays.toString(destinationsNames.toArray()));
+
+        //get specific address d2
+        Address res = kubernetes.getAddressClient(addressSpace.getMetadata().getNamespace()).withName(d2.getMetadata().getName()).get();
+        assertThat("Rest api does not return specific address", res.getSpec().getAddress(), is(d2.getSpec().getAddress()));
+
+        resourcesManager.deleteAddresses(d1);
+
+        //d2
+        response = AddressUtils.getAddresses(addressSpace).stream().map(address -> address.getSpec().getAddress()).collect(Collectors.toList());
+        assertThat("Rest api does not return right addresses", response, is(destinationsNames.subList(1, 2)));
+        LOGGER.info("address {} successfully deleted", d1.getSpec().getAddress());
+
+        resourcesManager.deleteAddresses(d2);
+
+        //empty
+        List<Address> listRes = AddressUtils.getAddresses(addressSpace);
+        assertThat("Rest api returns addresses", listRes, is(Collections.emptyList()));
+        LOGGER.info("addresses {} successfully deleted", d2.getSpec().getAddress());
+
+        resourcesManager.setAddresses(d1, d2);
+        resourcesManager.deleteAddresses(d1, d2);
+
+        listRes = AddressUtils.getAddresses(addressSpace);
+        assertThat("Rest api returns addresses", listRes, is(Collections.emptyList()));
+        LOGGER.info("addresses {} successfully deleted", Arrays.toString(destinationsNames.toArray()));
     }
 }
