+++ systemtests/src/test/java/io/enmasse/systemtest/iot/shared/control/CommandAndControlTest.java
@@ -11,7 +11,9 @@ import io.enmasse.systemtest.bases.iot.ITestIoTShared;
 import io.enmasse.systemtest.iot.CredentialsRegistryClient;
 import io.enmasse.systemtest.iot.DeviceRegistryClient;
 import io.enmasse.systemtest.iot.HttpAdapterClient;
+import io.enmasse.systemtest.iot.MessageType;
 import io.enmasse.systemtest.logs.CustomLogger;
+import io.enmasse.systemtest.utils.IoTUtils;
 import io.enmasse.systemtest.utils.TestUtils;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.json.JsonObject;
@@ -116,7 +118,7 @@ class CommandAndControlTest extends TestBase implements ITestIoTShared {
 
         var f1 = setupMessagingReceiver(sentFuture, null);
 
-        waitForFirstSuccessOnTelemetry(httpClient);
+        IoTUtils.waitForFirstSuccess(httpClient, MessageType.TELEMETRY);
 
         var response = sendTelemetryWithTtd();
 
@@ -144,7 +146,7 @@ class CommandAndControlTest extends TestBase implements ITestIoTShared {
             commandMessage.setReplyTo(replyToAddress);
         });
 
-        waitForFirstSuccessOnTelemetry(httpClient);
+        IoTUtils.waitForFirstSuccess(httpClient, MessageType.TELEMETRY);
 
         var response = sendTelemetryWithTtd();
